+++ src/main/java/ru/skypro/homework/service/impl/AuthServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
-import ru.skypro.homework.dto.Register;
+import ru.skypro.homework.dto.user.Register;
 import ru.skypro.homework.service.AuthService;
 
 @Service
