+++ src/main/resources/application.properties
@@ -1,0 +1,1 @@
+spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
